[refactor] Moved 'createAccordionElement' to its own TS file
